+++ packages/demo/src/routes/layout.tsx
@@ -106,9 +106,6 @@ const ROUTES = [
   "/subdir",
   "/subdir/other",
   "/error",
-  "/error-loader",
-  // TODO
-  // "/error-ssr",
 ];
 
 //
